App.js now in head and contains all minified required vendor libraries (jQuery, Foundation etc)
This is to allow executing jQuery $() ad-hoc, at any stage one the page. It might be needed when JS snippet needs to be added directly after some UI components (e.g. Slick Carousel), with that element data etc. In that case the only (known) solution is to invoke $() straight away.